Fix production value truncation in industry GUI GetIndustryString

+++ src/industry_gui.cpp
@@ -1625,7 +1625,7 @@ protected:
 			uint transported;     ///< Percent transported last month.
 			std::string suffix;   ///< Cargo suffix.
 
-			CargoInfo(CargoType cargo_type, uint16_t production, uint transported, std::string &&suffix) : cargo_type(cargo_type), production(production), transported(transported), suffix(std::move(suffix)) {}
+			CargoInfo(CargoType cargo_type, uint32_t production, uint transported, std::string &&suffix) : cargo_type(cargo_type), production(production), transported(transported), suffix(std::move(suffix)) {}
 		};
 		std::vector<CargoInfo> cargos;
